requirements.txt
annotated-types==0.7.0
anyio==4.12.1
argcomplete==3.6.3
asttokens==3.0.1
bitarray==3.8.0
black==26.1a1
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
ckzg==2.1.5
click==8.3.1
colorama==0.4.6
comm==0.2.3
contourpy==1.3.3
cryptography==46.0.3
cycler==0.12.1
cytoolz==1.1.0
datamodel-code-generator==0.52.2
debugpy==1.8.19
decorator==5.2.1
eth-account==0.13.7
eth-hash==0.7.1
eth-keyfile==0.8.1
eth-keys==0.7.0
eth-rlp==2.2.0
eth-typing==5.2.1
eth-utils==5.3.1
eth_abi==6.0.0b1
executing==2.2.1
finplot==1.9.7
fonttools==4.61.1
genson==1.3.0
h11==0.16.0
h2==4.3.0
hexbytes==1.3.1
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.11
inflect==7.5.0
iniconfig==2.3.0
ipykernel==7.1.0
ipython==9.9.0
ipython_pygments_lexers==1.1.1
isort==7.0.0
jedi==0.19.2
Jinja2==3.1.6
jupyter_client==8.8.0
jupyter_core==5.9.1
kiwisolver==1.4.9
MarkupSafe==3.0.3
matplotlib==3.10.8
matplotlib-inline==0.2.1
more-itertools==10.8.0
mypy_extensions==1.1.0
nest-asyncio==1.6.0
numpy==2.4.0
packaging==25.0
pandas==2.3.3
parsimonious==0.10.0
parso==0.8.5
pathspec==1.0.3
pexpect==4.9.0
pillow==12.1.0
platformdirs==4.5.1
pluggy==1.6.0
poly_eip712_structs==0.0.1
polymarket-gamma-sdk==1.0.0
prompt_toolkit==3.0.52
psutil==7.2.1
ptyprocess==0.7.0
pure_eval==0.2.3
py_builder_signing_sdk==0.0.2
py_clob_client==0.34.4
py_order_utils==0.3.2
pycparser==2.23
pycryptodome==3.23.0
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
pyparsing==3.3.1
PyQt6==6.10.2
PyQt6-Qt6==6.10.1
PyQt6_sip==13.11.0
pyqtgraph==0.14.0
pytest==9.0.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytokens==0.3.0
pytz==2025.2
PyYAML==6.0.3
pyzmq==27.1.0
qasync==0.28.0
regex==2025.11.3
requests==2.32.5
rlp==4.1.0
six==1.17.0
stack-data==0.6.3
toolz==1.1.0
tornado==6.5.4
tqdm==4.67.1
traitlets==5.14.3
typeguard==4.4.4
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
wcwidth==0.2.14
websockets==16.0
